Update web3j version. Update minor change in EmbeddedEthereum.kt

+++ src/main/kotlin/org/web3j/evm/EmbeddedEthereum.kt
@@ -256,7 +256,7 @@ class EmbeddedEthereum(
                 null, // TODO?
                 tcr.r,
                 tcr.s,
-                hexToULong(tcr.v).toInt(),
+                hexToULong(tcr.v),
                 tcr.type,
                 tcr.maxFeePerGas,
                 tcr.maxPriorityFeePerGas,
